+++ screens/Home.js
@@ -39,6 +39,23 @@ const Home = ({ navigation }) => {
     getPosts();
   }, []);
 
+  function timeConvert(time) {
+    time = time
+      .toString()
+      .match(/^([01]\d|2[0-3])(:)([0-5]\d)(:[0-5]\d)?$/) || [time];
+
+    if (time.length > 1) {
+      time = time.slice(1);
+      /* Show AM or PM based on time value */
+      time[5] = +time[0] < 12 ? "AM" : "PM";
+      /* Hide seconds */
+      time[3] = time[3] && "";
+      /* Convert from 24hr time to 12 hour time */
+      time[0] = +time[0] % 12 || 12;
+    }
+    return time.join("");
+  }
+
   const getPosts = async () => {
     setIsLoading(true);
     try {
@@ -141,7 +158,10 @@ const Home = ({ navigation }) => {
                           .replace("&#8211;", "-")
                           .replace("&#8217;", "'")}
                       </Text>
-                      <Text style={styles.date}>{item.date}</Text>
+                      <Text style={styles.date}>
+                        {item.date.replace(/-/g, "/").split("T")[0]}{" "}
+                        {timeConvert(item.date.split("T")[1])}
+                      </Text>
                     </View>
                     <RenderHtml
                       style={[styles.text]}